# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WAP — WordPress AI Platform
  version: 0.1.0
paths:
...
  /api/v1/chat/stream:
    post:
      summary: Chat Stream
      operationId: chat_stream_api_v1_chat_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/chat/{thread_id}/history:
    get:
      summary: Get History
      description: 'Return conversation history from LangGraph checkpoint.


        The thread_id must belong to the authenticated user to prevent data leakage.

        Accepts both the backend-internal format ({user_id}:{role}) and the PHP

        client conversationId format ({wp_user_id}:{product}:standard), translating

        the latter to the real internal thread_id via the session''s user_id.'
      operationId: get_history_api_v1_chat__thread_id__history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get History Api V1 Chat  Thread Id  History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatBody:
      properties:
        message:
          type: string
          minLength: 1
          title: Message
        page_context:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 64
          - type: 'null'
          title: Page Context
      type: object
      required:
      - message
      title: ChatBody
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer